ml/requirements.txt
# TomatoGuard/ml/requirements.txt - ML ONLY
keras==3.12.0
tensorflow==2.15.0
numpy==1.24.3
pandas==2.2.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0
opencv-python==4.8.1.78
tqdm==4.66.1